% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.evolution.r
\name{pop.evolution}
\alias{pop.evolution}
\title{Population evolution}
\usage{
pop.evolution(g0, e, r.ho, nu, N, nb.gen)
}
\arguments{
\item{g0}{Standard deviation of gametic value in a population without assortative mating}

\item{e}{Standard deviation of environmental effects}

\item{r.ho}{Correlation between mates}

\item{nu}{Correlation between parents' and offspring environmental effects}

\item{N}{Number of causal variants}

\item{nb.gen}{Number of generations}
}
\value{
A data frame with class "pop.evolution", with columns
't' (the generation, from 0 to nb.gen), 'N.kappa' (the value of \eqn{N \overline\kappa(t)}{N kappa bar(t)}),
'g' (the standard deviation of the gametic value, \eqn{g(t)}), 'r.ga' (the gametic correlation, \eqn{r_{ga}(t)}{r_ga(t)}), 
'rho' (the gene-environment correlation, \eqn{\rho(t)}{rho(t)}), 'a' (the standard deviation of the genetic value, \eqn{g(t)}),
and 'sigma2' (the variance of the phenotype, \eqn{sigma^2(t)}{sigma^2(t)}).

The parameters 'g0', 'e', 'r.ho', 'nu' and 'N' are stored in an attribute named 'parameters'.
}
\description{
Compute all parameters of the evolution of a population which is initially 
at gametic equilibrium, without assortative mating.
}
\details{
The details of the computation are given in the supplementary material of the paper,
in particular in appendix B.
}
\examples{
pop.evolution(1, 1, 0.6, 0.2, Inf, 10)

}
