% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.sim.r
\name{pop.sim}
\alias{pop.sim}
\title{Simple population simulations}
\usage{
pop.sim(
  g0,
  e,
  r.ho,
  nu,
  N = 100,
  nb.gen = 10,
  pop.size = 25000,
  digest = TRUE,
  keep.N.kappa = FALSE
)
}
\arguments{
\item{g0}{Standard deviation of gametic value in a population without assortative mating}

\item{e}{Standard deviation of environmental effects}

\item{r.ho}{Correlation between mates}

\item{nu}{Correlation between parents' and offspring environmental effects}

\item{N}{Number of causal variants}

\item{nb.gen}{Number of generations}

\item{pop.size}{Population size}

\item{digest}{Logical. Set to 'TRUE' to send back only a digest of the results}

\item{keep.N.kappa}{Logical. Set to 'TRUE' to keep trace of \eqn{N\overline\kappa(t)}{N kappa bar}.}
}
\value{
if 'digest' is 'TRUE', a data frame similar to the result of 'pop.evolution', with an additional
column for 'e' which contain the sd of the environmental components at each generation (this fluctuates
slightly around the value of the corresponding parameter, due to random sampling). If 'digest' is 'FALSE',
a list with member 'digest' (the digest as described before), 'G' (the matrix of genotypes at the last
generation), 'A' (the genetic value at the last generation), 'E' (the environmental value at the last 
generation), 'P' (the phenotypes at the last generation).
}
\description{
Simulate the evolution of a population with Assortative Mating and Vertical Cultural Transmission, with unlinked causal variants
}
\details{
Will simulate a population under Assortative Mating and Vertical Cultural Transmission with the parameters
supplied. Generations are non overlapping.

All SNPs unlinked. Their allele frequencies are drawn in a neutral site frequency spectrum, capped to [0.01, 0.99].
Allelic effect are drawn in a normal distribution, and normalized in order to get a standard deviation of the gametic value
equal to g0.
Mate pairs are formed and each pair has two offsprings.

Beware: setting 'keep.N.kappa' to 'TRUE' results in lengthy computations.
}
\examples{
# a simulation with N = 100 loci only
R <- pop.sim(g0 = sqrt(0.5), e = 1, r.ho = 0.6, nu = 0.4, pop.size = 25000, N = 100, nb.gen = 20, TRUE, TRUE)
# theoretical evolution
ev <- pop.evolution(g0 = sqrt(0.5), e = 1, r.ho = 0.6, nu = 0.4, N = 100, nb.gen = 20) 
# theoretical equilibrium values
limits <- AMVCT(g0 = sqrt(0.5), e = 1, r.ho = 0.6, nu = 0.4)

par(mfrow=c(2,2))
# plotting evolution of rho
plot(R$t, R$rho, type = "l", xlab = "t", ylab = expression(rho))
lines(ev$t, ev$rho, col = "red")
abline(h = limits$rho, col = "red", lty = 3)

# plotting evolution of N kappa bar
plot(R$t, R$N.kappa, type = "l", xlab = "t", ylab = expression(N * bar(kappa)))
lines(ev$t, ev$N.kappa, col = "red")
abline(h = 1/(1-limits$r.ga), col = "red", lty = 3)

# plotting evolution of a
plot(R$t, R$a, type = "l", xlab = "t", ylab = "a") 
lines(ev$t, ev$a, col = "red")
abline(h = limits$a, col = "red", lty = 3)

# plotting evolution of r.ga
plot(R$t, R$r.ga, type = "l", xlab = "t", ylab = expression(r[ga]))
lines(ev$t, ev$r.ga, col =" red")
abline(h = limits$r.ga, col = "red", lty = 3)

}
