% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.a.rho.r
\name{solve.a.rho}
\alias{solve.a.rho}
\title{Solve a non-linear system of equation for 'a' and 'rho'}
\usage{
\method{solve}{a.rho}(g0, e, r.AM, nu)
}
\arguments{
\item{g0}{Standard deviation of gametic value in a population without assortative mating}

\item{e}{Standard deviation of environmental effects}

\item{r.AM}{Correlation between mates (denoted \eqn{r_{ho}}{r_ho} in the paper)}

\item{nu}{Correlation between parents' and offspring environmental effects}
}
\value{
a vector with named components 'a' and 'rho'.
}
\description{
given the natural parameters \eqn{g_0}{g0}, \eqn{e}, \eqn{r_{AM}}{r_AM} and \eqn{\nu}{nu},
computes \eqn{a} and \eqn{rho}.
}
\details{
In the article text, 'r.AM' is denoted by \eqn{r_{ho}}{r_ho}.
The system formed by the equations 41 and 42 from the supplementary material is solved
to get the values of \eqn{a} and \eqn{\rho}{rho} at equilibrium.

The system may have solutions even if the inequality ensuring that the environmental
variance matrix of a trio is positive definite (equation 26). This inequality is tested, and
if it is not verified, a warning is emitted and NAs are returned.
}
